Move overlap logic to the thank you page
