+++ client/my-sites/checkout/checkout-thank-you/akismet-checkout-thank-you.tsx
@@ -1,11 +1,13 @@
+import { AKISMET_PRODUCTS_LIST, isAkismetProduct } from '@automattic/calypso-products';
 import { Button, Card } from '@automattic/components';
 import { createInterpolateElement } from '@wordpress/element';
 import { sprintf, __ } from '@wordpress/i18n';
 import classNames from 'classnames';
 import { useTranslate } from 'i18n-calypso';
-import { useEffect, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import QueryProducts from 'calypso/components/data/query-products-list';
+import QueryUserPurchases from 'calypso/components/data/query-user-purchases';
 import ClipboardButton from 'calypso/components/forms/clipboard-button';
 import FormTextInput from 'calypso/components/forms/form-text-input';
 import Main from 'calypso/components/main';
@@ -13,6 +15,7 @@ import useAkismetKeyQuery from 'calypso/data/akismet/use-akismet-key-query';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { isProductsListFetching, getProductName } from 'calypso/state/products-list/selectors';
+import { getUserPurchases, isFetchingUserPurchases } from 'calypso/state/purchases/selectors';
 import type { FunctionComponent } from 'react';
 
 interface AkismetCheckoutThankYouProps {
@@ -27,7 +30,9 @@ const AkismetCheckoutThankYou: FunctionComponent< AkismetCheckoutThankYouProps >
 	const productName = useSelector( ( state ) =>
 		hasProduct ? getProductName( state, productSlug ) : null
 	);
-	const isLoading = useSelector( isProductsListFetching );
+	const isLoading = useSelector(
+		( state ) => isProductsListFetching( state ) || isFetchingUserPurchases( state )
+	);
 
 	const onManagePurchaseClick = () => {
 		dispatch(
@@ -36,6 +41,37 @@ const AkismetCheckoutThankYou: FunctionComponent< AkismetCheckoutThankYouProps >
 			} )
 		);
 	};
+
+	const userActivePurchases = useSelector(
+		( state ) => getUserPurchases( state )?.filter( ( purchase ) => purchase.active ) ?? []
+	);
+
+	const akismetPurchaseThatOverlapsPurchasedProduct = useMemo( () => {
+		const akismetPurchases = userActivePurchases.filter(
+			( purchase ) => isAkismetProduct( purchase ) && 'siteless.akismet.com' === purchase.domain
+		);
+		const akismetProducts = AKISMET_PRODUCTS_LIST as ReadonlyArray< string >;
+
+		const lowerTierProducts = akismetPurchases.filter(
+			( { productSlug: ps } ) =>
+				akismetProducts.indexOf( productSlug ) > akismetProducts.indexOf( ps )
+		);
+
+		if ( lowerTierProducts.length === 0 ) {
+			return null;
+		}
+
+		const highestTierOverlappedProduct = lowerTierProducts[ lowerTierProducts.length - 1 ];
+
+		// If the highest tiered product is Akismet Free, return null
+		// We don't need to display a notice about multiple paid plans in this case
+		if ( highestTierOverlappedProduct.productSlug === 'ak_free_yearly' ) {
+			return null;
+		}
+
+		// Returns the highest tiered product
+		return lowerTierProducts[ lowerTierProducts.length - 1 ];
+	}, [ productSlug, userActivePurchases ] );
 
 	return (
 		<Main className="akismet-checkout-thank-you">
@@ -49,6 +85,7 @@ const AkismetCheckoutThankYou: FunctionComponent< AkismetCheckoutThankYouProps >
 			<Card className="akismet-checkout-thank-you__card">
 				{ /* This product query takes a while to load, an improvment here would be to add a type filter on wpcom like jetpack has. See: 2f832-pb/ */ }
 				{ hasProduct && <QueryProducts /> }
+				<QueryUserPurchases />
 
 				<h2 className="akismet-checkout-thank-you__main-message">
 					{ /* the single space literal below is intentional for rendering purposes */ }
@@ -57,6 +94,14 @@ const AkismetCheckoutThankYou: FunctionComponent< AkismetCheckoutThankYouProps >
 						{ String.fromCodePoint( 0x1f389 ) /* Celebration emoji 🎉 */ }
 					</span>
 				</h2>
+
+				{ akismetPurchaseThatOverlapsPurchasedProduct && (
+					<div className="akismet-checkout-thank-you__overlap-notice">
+						<p>
+							Thanks for upgrading! Be sure to cancel any overlapping purchases in your account.
+						</p>
+					</div>
+				) }
 
 				<p
 					className={
