+++ client/my-sites/checkout/composite-checkout/components/prepurchase-notices/index.jsx
@@ -6,15 +6,12 @@ import {
 	isJetpackScanSlug,
 	getAllFeaturesForPlan,
 	planHasSuperiorFeature,
-	isAkismetProduct,
-	AKISMET_PRODUCTS_LIST,
 } from '@automattic/calypso-products';
 import { useShoppingCart } from '@automattic/shopping-cart';
 import { useEffect, useMemo } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import Notice from 'calypso/components/notice';
 import useCartKey from 'calypso/my-sites/checkout/use-cart-key';
-import { getUserPurchases } from 'calypso/state/purchases/selectors';
 import { requestRewindCapabilities } from 'calypso/state/rewind/capabilities/actions';
 import {
 	getSitePlan,
@@ -23,7 +20,6 @@ import {
 	getSiteOption,
 } from 'calypso/state/sites/selectors';
 import getSelectedSite from 'calypso/state/ui/selectors/get-selected-site';
-import AkismetProductOverlapsOwnedProductNotice from './akismet-product-overlaps-owned-product-notice';
 import CartPlanOverlapsOwnedProductNotice from './cart-plan-overlaps-owned-product-notice';
 import JetpackPluginRequiredVersionNotice from './jetpack-plugin-required-version-notice';
 import SitePlanIncludesCartProductNotice from './site-plan-includes-cart-product-notice';
@@ -37,9 +33,6 @@ const PrePurchaseNotices = () => {
 	const dispatch = useDispatch();
 
 	const selectedSite = useSelector( getSelectedSite );
-	const userActivePurchases = useSelector(
-		( state ) => getUserPurchases( state )?.filter( ( purchase ) => purchase.active ) ?? []
-	);
 	const siteId = selectedSite?.ID;
 	const cartKey = useCartKey();
 	const { responseCart } = useShoppingCart( cartKey );
@@ -87,41 +80,6 @@ const PrePurchaseNotices = () => {
 		} );
 	};
 
-	const akismetPurchaseThatOverlapsCartProduct = useMemo( () => {
-		const productSlugInCart = cartItemSlugs.find( ( productSlug ) =>
-			isAkismetProduct( { productSlug } )
-		);
-		const akismetPurchases = userActivePurchases.filter(
-			( purchase ) => isAkismetProduct( purchase ) && 'siteless.akismet.com' === purchase.domain
-		);
-
-		// Continue only if the cart includes Akismet product
-		if ( ! productSlugInCart ) {
-			return null;
-		}
-
-		const lowerTierProducts = akismetPurchases.filter(
-			( { productSlug } ) =>
-				AKISMET_PRODUCTS_LIST.indexOf( productSlugInCart ) >
-				AKISMET_PRODUCTS_LIST.indexOf( productSlug )
-		);
-
-		if ( lowerTierProducts.length === 0 ) {
-			return null;
-		}
-
-		const highestTierOverlappedProduct = lowerTierProducts[ lowerTierProducts.length - 1 ];
-
-		// If the highest tiered product is Akismet Free, return null
-		// We don't need to display a notice about multiple paid plans in this case
-		if ( highestTierOverlappedProduct.productSlug === 'ak_free_yearly' ) {
-			return null;
-		}
-
-		// Returns the highest tiered product
-		return lowerTierProducts[ lowerTierProducts.length - 1 ];
-	}, [ cartItemSlugs, userActivePurchases ] );
-
 	const siteProductThatOverlapsCartPlan = useMemo( () => {
 		const planSlugInCart = cartItemSlugs.find( isJetpackPlanSlug );
 		if ( ! planSlugInCart || ! currentSiteProducts ) {
@@ -157,14 +115,6 @@ const PrePurchaseNotices = () => {
 			backupPluginActive || isJetpackMinimumVersion( state, siteId, BACKUP_MINIMUM_JETPACK_VERSION )
 		);
 	} );
-
-	if ( akismetPurchaseThatOverlapsCartProduct ) {
-		return (
-			<AkismetProductOverlapsOwnedProductNotice
-				purchase={ akismetPurchaseThatOverlapsCartProduct }
-			/>
-		);
-	}
 
 	// All these notices (and the selectors that drive them)
 	// require a site ID to work. We should *conceptually* always
